scripts/requirements.txt
beautifulsoup4
Pillow
